+++ src/renderer/components/workbench/workbench-settings-overlay.service-tier.test.tsx
@@ -62,10 +62,6 @@ function renderOverlay() {
         sidebarTopLevelSectionOrder={["recents", "cards", "threads", "files"]}
         sidebarTopLevelSections={makeDefaultSidebarTopLevelSectionsPrefs()}
         onSidebarTopLevelSectionVisibleChange={() => {}}
-        stageRailLayoutMode="sliding-window"
-        onStageRailLayoutModeChange={() => {}}
-        nextPanelPeekPx={32}
-        onNextPanelPeekPxChange={() => {}}
         threadQueueFollowUpsEnabled={false}
         onThreadQueueFollowUpsEnabledChange={() => {}}
         composerEnterBehavior="enter"